Run release only when main is change.
The release CI job should run only when a change in the main branch is
pushed. Thus, this commit removes the trigger used to launch this job
when the "update chart" job finished. This was originally added to fully
automate the release process with no PR. However, the current approach
creates a PR to be approved. Thus, this trigger can be removed.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>