Update kubewarden components hauler (#149)
* refactor(update-deps): dynamically generate updatecli pipeline.

Updates the updatecli pipeline to generate the sources and targets to
updates OCI artifact versions based on the configuration values.
Therefore, when a new dependency is added, only the values files needs
to be updated.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(update-deps): update Hauler manifest.

Updates the updatecli script used to update the policies and kuberlr
version to also update the version in the Hauler manifest.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(dep): automate policy-reporter update.

Changes the updatecli pipeline used to update dependencies to also
updates policy-reporter dependency resources like Helm chart and
container images.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(update-hauler): new updatecli policy to update Hauler manifest.

Add a new policy which extract the OCI artifacts and Helm chart version
from the Helm charts and update the Hauler manifest with them.

With the new updatecli policy used exclusively used to update the Hauler
manifest based on the values defined in the Helm chart, the update-deps
policy does not need to update the Hauler manifest. The code for that
has been moved to the new policy. update-deps can focus on the Helm
chart update only.

In a future change both policies will be used together to update every
that is required.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): renamte update-deps.yaml to update-chart-deps.yaml.

Rename the policy used to update the Helm chart dependencies from
update-deps.yaml to update-chart-deps.yaml. This is usefull because in a
future change this policy will be called together with the other
policies in a updatecli compose. Which can cause some names confusion.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(updatecli): updatecli compose to update Helm charts.

Adds bunch of updatecli compose files to update dependencies, updates
the Kubewarden versions in the Helm charts. As well as update the Hauler
manifest with the latest versions. There are compose files for all the
update types in use.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(ci): update dependencies and Hauler when updating charts.

Changes the CI job used to update the Helm chart to use the updatecli
compose file which will update third party dependencies, policies and
the Kuberwarden components versions and Hauler manifest. Previous this
change, the job was only updating the Kubewarden components.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* feat(ci): Upload Hauler manifest as a release asset

This change updates the CI job for the Helm chart release, adding the
Hauler manifest as a release asset to the GitHub release page.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): use pipelineid.

Updates the policies and the values file to define the pipelineid if the
values file has the pipelineid field. This allow the compose file create
a single PR for all the policies it run. Intead of having individual PR
for each policy.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): read file from updatecli branch.

When updating the Hauler manifest the updatecli policy should read the
Helm chart files from the branch under manipulation. Therefore, changes
from other policies run in the compose file can be considered in the
manifest update like Helm chart version and Kubewarden policies version
changes.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): PR title value.

Adds a new value field allowing customization of the PR title open by
the compose file.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): add values for release jobs.

Adds an updatecli value file to keep the values used only by the release
jobs.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

* fix(updatecli): compose file to update dependencies.

Adds an updatecli compose file to update the Helm chart dependencies and
Hauler manifest in a single run.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>

---------

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>